Added files via upload
Registration vector calculation on page 5. The only thing I haven't implemented is the last step (described at the top of the page, q = [qr|qt]t) just because I don't know what that | operator is and a google search didn't turn up much - anybody know this? So right now it's returning qr and qt, the optimal rotation and translation vectors. Also the optimal translation vector qt is coming out as a 3x3 matrix, not a 1d vector, and I'm not sure if that's right or wrong. Lemme know how this looks / if this makes sense